use dayobs, no of nights and instrument  as URL parameters to use in other child routes and rerender the child route upon changing the parameters
