+++ src/pages/data-log.jsx
@@ -1,9 +1,11 @@
 import React from "react";
-import { useRouter } from "@tanstack/react-router";
+// import { useRouter } from "@tanstack/react-router";
+import { useSearch } from "@tanstack/react-router";
 
 function DataLog({ loading = false }) {
-  const { params } = useRouter().options.context;
-  const { dayobs, noOfNights, instrument } = params;
+  // const { params } = useRouter().options.context;
+  // const { dayobs, noOfNights, instrument } = params;
+  const { dayObs, noOfNights, instrument } = useSearch({ from: "__root__" });
 
   return (
     <div>
@@ -16,7 +18,7 @@ function DataLog({ loading = false }) {
             This is a placeholder for the data log.
           </div>
           <div className="text-sm text-gray-500">
-            Day Observed: {dayobs.toLocaleDateString()}
+            Day Observed: {dayObs}
             <br />
             Number of Nights: {noOfNights}
             <br />
