+++ src/pages/summary.jsx
@@ -1,0 +1,280 @@
+import { useState, useEffect } from "react";
+import { Toaster } from "@/components/ui/sonner";
+import { toast } from "sonner";
+import Applet from "@/components/applet.jsx";
+import AppletExposures from "@/components/applet-exposures.jsx";
+import MetricsCard from "@/components/metrics-card.jsx";
+import { DateTime } from "luxon";
+
+import { EfficiencyChart } from "@/components/ui/RadialChart.jsx";
+import ShutterIcon from "../assets/ShutterIcon.svg";
+import TimeLossIcon from "../assets/TimeLossIcon.svg";
+import JiraIcon from "../assets/JiraIcon.svg";
+import {
+  fetchExposures,
+  fetchAlmanac,
+  fetchNarrativeLog,
+  fetchExposureFlags,
+  fetchJiraTickets,
+} from "@/utils/fetchUtils";
+import {
+  calculateEfficiency,
+  calculateTimeLoss,
+  getDayobsStr,
+  getDatetimeFromDayobsStr,
+} from "@/utils/utils";
+import DialogMetricsCard from "@/components/dialog-metrics-card";
+import JiraTicketsTable from "@/components/jira-tickets-table";
+import { useSearch } from "@tanstack/react-router";
+
+export default function Summary() {
+  // const router = useRouter();
+  const { dayObs, noOfNights, instrument } = useSearch({ from: "__root__" });
+  const dayobs = DateTime.fromFormat(dayObs.toString(), "yyyyLLdd").toJSDate();
+  const [nightHours, setNightHours] = useState(0.0);
+  const [weatherLoss, setWeatherLoss] = useState(0.0);
+  const [faultLoss, setFaultLoss] = useState(0.0);
+  const [exposureFields, setExposureFields] = useState([]);
+  const [exposureCount, setExposureCount] = useState(0);
+  const [sumExpTime, setSumExpTime] = useState(0.0);
+  const [flags, setFlags] = useState([]);
+
+  const [exposuresLoading, setExposuresLoading] = useState(true);
+  const [almanacLoading, setAlmanacLoading] = useState(true);
+  const [narrativeLoading, setNarrativeLoading] = useState(true);
+
+  const [jiraTickets, setJiraTickets] = useState([]);
+  const [jiraLoading, setJiraLoading] = useState(true);
+
+  const [flagsLoading, setFlagsLoading] = useState(true);
+
+  // const handleDayobsChange = (date) => {
+  //   setDayobs(date);
+  // };
+
+  // const handleNoOfNightsChange = (nightsCount) => {
+  //   setNoOfNights(nightsCount);
+  // };
+
+  // const handleInstrumentChange = (inst) => {
+  //   setInstrument(inst);
+  // };
+
+  useEffect(() => {
+    const abortController = new AbortController();
+
+    setExposuresLoading(true);
+    setAlmanacLoading(true);
+    setNarrativeLoading(true);
+    setJiraLoading(true);
+    setFlagsLoading(true);
+
+    if (!dayObs || isNaN(dayObs)) {
+      toast.error("Invalid Date Selected! Please select a valid date.");
+    }
+    // const dayobs = DateTime.fromFormat(dayObs.toString(), "yyyyLLdd");
+    // const dayobsStr = `${dayObs}`;
+    let dayobsStr = getDayobsStr(dayobs);
+    if (!dayobsStr) {
+      toast.error("No Date Selected! Please select a valid date.");
+      setExposuresLoading(false);
+      setAlmanacLoading(false);
+      setNarrativeLoading(false);
+      setJiraLoading(false);
+      setFlagsLoading(false);
+      return;
+    }
+    const dateFromDayobs = getDatetimeFromDayobsStr(dayobsStr);
+    const startDate = dateFromDayobs.minus({ days: noOfNights - 1 });
+    const startDayobs = startDate.toFormat("yyyyLLdd");
+    const endDate = dateFromDayobs.plus({ days: 1 });
+    const endDayobs = endDate.toFormat("yyyyLLdd");
+
+    fetchExposures(startDayobs, endDayobs, instrument, abortController)
+      .then(([exposureFields, exposuresNo, exposureTime]) => {
+        setExposureFields(exposureFields);
+        setExposureCount(exposuresNo);
+        setSumExpTime(exposureTime);
+        setExposuresLoading(false);
+        if (exposuresNo === 0) {
+          toast.warning("No exposures found for the selected date range.");
+        }
+      })
+      .catch((err) => {
+        if (!abortController.signal.aborted) {
+          const msg = err?.message;
+          toast.error("Error fetching exposures!", {
+            description: msg,
+            duration: Infinity,
+          });
+        }
+      })
+      .finally(() => {
+        if (!abortController.signal.aborted) {
+          setExposuresLoading(false);
+        }
+      });
+
+    fetchAlmanac(startDayobs, endDayobs, abortController)
+      .then((hours) => {
+        setNightHours(hours);
+      })
+      .catch((err) => {
+        if (!abortController.signal.aborted) {
+          const msg = err?.message;
+          toast.error("Error fetching almanac!", {
+            description: msg,
+            duration: Infinity,
+          });
+        }
+      })
+      .finally(() => {
+        if (!abortController.signal.aborted) {
+          setAlmanacLoading(false);
+        }
+      });
+
+    fetchNarrativeLog(startDayobs, endDayobs, instrument, abortController)
+      .then(([weather, fault]) => {
+        setWeatherLoss(weather);
+        setFaultLoss(fault);
+        if (weather === 0 && fault === 0) {
+          toast.warning("No time loss reported in the Narrative Log.");
+        }
+      })
+      .catch((err) => {
+        if (!abortController.signal.aborted) {
+          const msg = err?.message;
+          toast.error("Error fetching narrative log!", {
+            description: msg,
+            duration: Infinity,
+          });
+        }
+      })
+      .finally(() => {
+        if (!abortController.signal.aborted) {
+          setNarrativeLoading(false);
+        }
+      });
+
+    fetchJiraTickets(startDayobs, endDayobs, instrument, abortController)
+      .then((issues) => {
+        setJiraTickets(issues);
+        if (issues.length === 0) {
+          toast.warning("No Jira tickets reported.");
+        }
+      })
+      .catch((err) => {
+        if (!abortController.signal.aborted) {
+          const msg = err?.message;
+          toast.error("Error fetching Jira!", {
+            description: msg,
+            duration: Infinity,
+          });
+        }
+      })
+      .finally(() => {
+        if (!abortController.signal.aborted) {
+          setJiraLoading(false);
+        }
+      });
+
+    fetchExposureFlags(startDayobs, endDayobs, instrument, abortController)
+      .then((flags) => {
+        setFlags(flags);
+        if (flags.length === 0) {
+          toast.warning("No exposures flagged for the selected date range.");
+        }
+      })
+      .catch((err) => {
+        if (!abortController.signal.aborted) {
+          const msg = err?.message;
+          toast.error("Error fetching exposure flags!", {
+            description: msg,
+            duration: Infinity,
+          });
+        }
+      })
+      .finally(() => {
+        if (!abortController.signal.aborted) {
+          setFlagsLoading(false);
+        }
+      });
+
+    return () => {
+      abortController.abort();
+    };
+  }, [dayObs, dayobs, noOfNights, instrument]);
+
+  // calculate open shutter efficiency
+  const efficiency = calculateEfficiency(nightHours, sumExpTime, weatherLoss);
+  const efficiencyText = `${efficiency} %`;
+  const [timeLoss, timeLossDetails] = calculateTimeLoss(weatherLoss, faultLoss);
+  const newTicketsCount = jiraTickets.filter((tix) => tix.isNew).length;
+
+  return (
+    <>
+      <div className="flex flex-col w-full p-8 gap-8">
+        {/* Metrics */}
+        <div className="grid grid-cols-1 sm:grid-cols-2 lg:grid-cols-4 gap-8">
+          <MetricsCard
+            icon={ShutterIcon}
+            data={exposureCount}
+            label="Nighttime exposures taken"
+            metadata="(TBD expected)"
+            tooltip="On-sky exposures taken during the specified date range."
+            loading={exposuresLoading}
+          />
+          <MetricsCard
+            icon={<EfficiencyChart value={efficiency} />}
+            data={efficiencyText}
+            label="Open-shutter (-weather) efficiency"
+            tooltip="Efficiency computed as total exposure time / (time between 18 degree twilights minus time lost to weather)"
+            loading={almanacLoading || exposuresLoading}
+          />
+          <MetricsCard
+            icon={TimeLossIcon}
+            data={timeLoss}
+            label="Time loss"
+            metadata={timeLossDetails}
+            tooltip="Time loss as reported in the Narrative Log."
+            loading={narrativeLoading}
+          />
+          <DialogMetricsCard
+            icon={JiraIcon}
+            data={newTicketsCount}
+            label="Jira tickets created"
+            metadata={`(${jiraTickets.length - newTicketsCount} updated)`}
+            tooltip="Jira tickets created or updated within the specified date range."
+            loading={jiraLoading}
+            dialogTitle="Jira Tickets"
+            dialogDescription="List of Jira tickets created or updated within the specified date range."
+            dialogContent={
+              <JiraTicketsTable loading={jiraLoading} tickets={jiraTickets} />
+            }
+          ></DialogMetricsCard>
+        </div>
+        {/* Applets */}
+        <div className="flex flex-col gap-4">
+          <div className="grid grid-cols-1 lg:grid-cols-2 gap-4">
+            <AppletExposures
+              exposureFields={exposureFields}
+              exposureCount={exposureCount}
+              sumExpTime={sumExpTime}
+              flags={flags}
+              exposuresLoading={exposuresLoading}
+              flagsLoading={flagsLoading}
+            />
+            <Applet />
+          </div>
+          <div className="grid grid-cols-1 lg:grid-cols-3 gap-4">
+            <Applet />
+            <Applet />
+            <Applet />
+          </div>
+        </div>
+      </div>
+      <Toaster expand={true} richColors closeButton />
+    </>
+  );
+}
